requirements.txt
Django==5.2.2
djangorestframework==3.16.0
Pillow==11.3.0
django-filter==24.2

